virtualrobot_service/requirements.txt
Django==5.2
djangorestframework==3.16.0
tensorflow    # để load .h5
numpy
